render: Move some loose utility functions to utils.rs in wgpu renderer
